Update matrix to remove missing connections
The switch side of lefthand diodes no longer need to connect righthand diodes. Diodes join to the same row trace, but the switch-to-diode connections are separate (except for the switches that overlap with nav. switches — these are joined to reduct routes). The switches in the schematic are now represented as SPDT switches, which have two outputs instead of one. The footprints have been updated to number the righthand side leg to pin 3 (instead of being another pin 1).